Improve Sandbox handling of transitions list
Sandbox will check first all components to find one that can simulate the transition incoming.
Otherwise, it will use the first component that can process the transition generic incoming.
If the incoming list, the handling will assume that the transition is the root of a TransitionTree.
A parameter is added to allow to execute transition that is not from a Component's state machine.

--HG--
branch : storage