Add issue reporting mechanism.
Two decorators are available from checkmate._issue:
  - report: to report an open issue
  - fix: to close the issue and create a regression test.

The reported issue will be appended to the method docstring and be used when running doctest.
Running the reported issue with @report() will fail if the doctest succeed.
Running the reported fix with @fix() will succeed if the doctest fails.

When the issue is fixed, the issue should be checked against regression by changing @report() into @fix() in the code.

--HG--
branch : issue_report