Improve the InternalStorage factory do handle final state better
The change from class to __init__ method is move out of factory.
This is now done at TransitionStorage creation stage for final state storage only.
The factory code is now extended to handle state instance argument.
It supports both instance to the __init__() and the toggle(), ...

The sandbox is changed to feed the procedure with its final state.

--HG--
branch : compare_final