Add default keyword only argument to Partition
If set to False, the Partition will constructed without default value.
The value of the resulting partition will be None.
All attributes will have value set to None too.
If set to True, the default value will be the first valid value from yaml.

The default value of the default argument is True.

--HG--
branch : devel